move flush to flush.go
